refactor(mine_loop): Use channel for preprocess task return value
There are two options for getting the result of the preprocess task to the mine
loop: via return value (up until now) and via a channel (from now on). A channel
is preferable because the other sub-{task,loop}s of the mine loop also use
channels. As a result, the mine loop has uniform push-based control flow. The
alternative would be to mix paradigms: push-notifications from channels with
polling of task handles to see if they are finished. (Not good.)