refactor(mine_loop): Allow mining without peers if not on main net
Mining without peers is disabled in this application since we don't want
users to waste energy unless they're reasonably sure that they are on
the tip. Checking if they have any peers is a very good heuristic for
catching the negative case. However, to make some
integrations/end-to-end tests easier to perform, without needing
connected peers, this commit limits this rule to only apply on main net.
In other words: On any other network than main, you can now mine without
having any peers.

This rule is specifically intended to facilitate an end-to-end test
on the testnet network that the node's software handles both mining and
verification correctly across the upcoming alpha hardfork. The correct
transition across a hardfork can then be tested by resetting the block
to hard fork height minus two (118 on testnet, as hardfork occurs at 120
on testnet), and then starting the client with both guessing and
composition active.