fix(pow): Ensure indices cannot be reused across proposals after HF-alpha
In ConsensusRuleSet::Reboot the opened indices depends on the concrete
proposal being guessed on. But because of a bug in
85e6799c116f1f9082720e6d250c45620a0f6b78 and in
8410301581cdcb793624c099b27a9f76531cc88b, this was not the case on
HardFork-alpha. Concretely an opened Merkle tree authentication path
could be used for as many block proposals as was known. This would have
been very bad for transaction throughput since the solution would have
been to an arbitrary proposals with the right parent. With this fix, the
indices of the opened leafs depend on the concrete proposal. But the
Merkle tree stays unchanged, it still only depends on the parent block's
hash.

See #727:
https://github.com/Neptune-Crypto/neptune-core/pull/727#issuecomment-3435837295

Add test: Ensure that indices cannot be reused over multiple proposals.