refactor(rapid-block-download): Accommodate flaky peers
When peers are flaky, they might not respond to block requests. This commit
hardens the asynchronous architecture of the sync loop to account for peers that
can be flaky. Specifically:

 1. Prefer `try_send` over `send` when sending messages from the sync loop to
    the main loop.

The latter blocks if and as long as the channel capacity is maxed out. If the
counterpart also blocks (for the exact same reason or any othe reason), then
you have a deadlock. In contrast, `try_send` returns immediately and yields an
error if the transmission was not successful.

 2. Factor out the function that sends tip-successors to the main loop in order
    into an asynchronous task.

This function might take a long while to finish, particularly if it spends time
waiting on `send` to complete. And in this particular case we definitely want to
use `send` rather than `try_send` because we want to ensure that these blocks
arrive at the main loop in good order.