feat(rapid-block-download): Resume sync from stored state
When syncing, downloaded blocks are stored to a temp directory. If the node
crashes or even is powered down gracefully, those files might still be there.
So upon (re-)booting the sync loop, check if there are any blocks already
in the temp directory so that we can avoid downloading them again.