refactor(rapid-block-download): Drop functionality for sanctioning peers
There is a potentially very large temporal distance between the peer's behavior
(sending the block) and judgment (block was valid or invalid). In order to
punish or reward peers appropriately, we would have to carry around this
information (who sent the block) until we can act on it (by punishing or
rewarding them). More information carried across time increases the complexity
of an already complex subsystem.