fix: Make test case generator agree across Linux/MacOS
Although a custom hash map implementation that was supposed to use
deterministic iterators was used here, this determinism did not agree:
On Linux and Windows, one result was returned and on MacOS another. The
solution is to simply add a sorting of keys of a hash map before
collecting the keys into a vector.

This closes #760.