feat(rapid-block-download): Sync from dynamic peer set
Enable two message types to the sync loop, for adding and removing peers. Then
the sync loop has to sync from whatever peers are available.

To avoid a situation where one loop sends more messages than it receives for a
sustained period of time, the code is adjusted in two ways.

 1. The block requests are not sent out individually but collectively in a
    batch.

 2. Transmissions of block requests and block responses are sent without
    blocking and are allowed to fail if the channel is out of capacity. All
    other message types are sent in a blocking way ensuring that they are
    received in good order.