fix(rapid-block-download): Wait for subtask before terminating sync loop
There is pathway to escape the sync loop when for a certain period no blocks
have come in. Previously, that would trigger the termination of the sync loop.
Now, that's only true if the tip-successors subtask is not active. So if it is,
we wait for it to finish first.

Also:
 - Figure out early in the pipeline that we do not need to send out block
   requests because all blocks are in already.
 - Be verbose about failures to send tip-successors to the main loop before
   retrying.