perf(rapid-block-download): Prefer `VecDeque` for `pop_front`
When syncing progresses in good order and the first 32 or so blocks are
processed, we need to get rid of the first limb in the synchronization bit mask.
Previously, this mask stored a `Vec` of `u32`s and would `remove(0)` when
necessary. Now it is a `VecDeque` instead, which comes with the much cheaper
`pop_front()`.