refactor(rapid-block-download): Validate blocks prior to sync->main transmission
When the sync loop receives blocks, it has no way of verifying their validity.
The main loop expects incoming blocks to be valid (ignoring blocks forwarded
from a peer to the sync loop). Therefore, the sync loop should verify the blocks
prior to sending them in order to the main loop.

To facilitate this check, this commit introduces a new object, `BlockValidator`,
which contains the function that tests blocks for validity. In production, this
object always wrap the canonical block validation function. But in test mode,
the test setup may inject a bogus validation function. Moreover, it means that
the sync loop can make abstraction over block validation logic.

Since blocks are validated relative to their predecessors, the sync loop now
stores the current tip (previously it was only the current tip height). The
alternative would be to fetch it from the main loop or pass a mutex. Whenever
the sync loop is running, the main loop is only allowed to set the tip on the
orders of the sync loop. So in particular the sync loop always knows what the
tip is by just remembering its most recent order. There is no need for state
reconciliation this way.