refactor(peer_loop): Modify height-tracker only if block handling success
Struct `MutablePeerState`, of which one object lives in each per loop by the
name `peer_state_info`, has a field `highest_shared_block_height`. It is used
for two things.

 1. In the shallow fork reconciliation mechanism, AKA. `try_ensure_path`, after
    successfully passing a path of blocks to the main loop, control must make a
    decision whether to send a new `BlockHeightRequest` to the peer in question.
    It makes this decision based on whether the biggest block height in the path
    of blocks passed to the main loop is smaller than the peer's
    `highest_shared_block_height`, which could have been set in the mean time by
    a `BlockNotification`.
 2. When the peer loop receives a `Block` from the main loop, control needs to
    decide whether to send a `BlockNotification` to said peer. It makes this
    decision based on whether the new block's height is larger than the peer's
    `highest_shared_block_height`.

Until now, it was set in two circumstances.

 1. If the peer is notified of the `Block` that just came in from the main loop.
 2. The peer sends a `BlockNotification`.
 3. When a `Block` from a peer is received, and only if no shallow fork
    reconciliation is running.

This commit changes (3) into

 3'. When a `Block` from a peer is received, leading to the completion of a
     shallow fork reconciliation process (which could be trivial if we are only
     one block behind) which itself results in one `Block` or batch of `Block`s
     being sent to the main loop for processing. In this case, the
     `highest_shared_block_height` is set to the largest height of all blocks
     thusly sent to the main loop.

This change anticipates interfacing with the sync loop, which will request
blocks out of any order. When those blocks come in, the
`highest_shared_block_height` should *not* be set. Moreover, when those blocks
come in, the shallow fork reconciliation mechanism should bot be activated
either. So it makes sense to tie these two non-events together.

The return type of `try_ensure_path` was modified to communicate the potential
new highest shared block height to the caller, which is the peer event loop.
This way, all modifications to `highest_shared_block_height` happen in this
loop (besides initialization, where the value is obtained from the handhsake).