refactor(sync-challenge): Increase timeout from 45s to 90s
Two out of three bootstrapper nodes listed in the README.md consistently respond
to the sync challenge after the timeout period has expired. The third one (which
consequently carries all the load for syncing nodes) responds on time *some of
the time*. So for whatever, the sync challenge response is hard to produce
and/or deliver.

This commit doubles the timeout period.