refactor(rapid-block-download): Relay sync blocks to sync loop
Modify the peer loop: upon receiving a block, check if we are in sync
mode. If so, then (depending on the block's features) relay it to the
main loop without validation, so that it relays it to the sync loop.

Modify the channel from peer loop to main loop: add two message types,
one for a new sync target (which extends the chain) and one for a new
block (living somewhere in the middle of the chain). Also: put the
data structures specific to main loop <-> peer loop interaction into a
separate file.

Modify main loop: add optional handle on sync loop to state, and handle
the new messages by relaying them if the handle is present.