ux!: Report sync status with more precision
Specifically, modify dashboard overview data field `syncing` into `sync_status`
which is a new enum where some variants have associated data. Render this field
correctly in the dashboard. The RPC server obtains the value of this field from
network state. It is updated by the main loop or peer loops. To get the progress
indicator, a message must be sent to the sync loop. The RPC handler sends a new
message to the main loop, asking it to update status indicators. The main loop's
handler for this message sends a message to the sync loop (if any) to compute
sync progress.