Working: DbtMap without `remove` functionality
Intended for a new table in the wallet database to allow for a fast
lookup from AOCL leaf index to MonitoredUtxo. The `MonitoredUtxo`s are
stored in a `DbtVec` list which is indexed by `u64`, and the AOCL is
also indexed by `u64`. This means that the motivation for and use case
I currently have in mind is `DbtMap<u64, Vec<u64>>`, where the inner vec
will might multiple entries in the case of reorganizations.

The reason that the `DbtMap` does not have remove functionality is that
we want the option of iterating through all keys. So all keys are stored
in an underlying `DbtVec` which only supports `pop`, `push`, and `set`
operations. Since an arbitrary element cannot be efficiently removed
from `DbtVec` this means that the same slow running time would apply for
`remove` on `DbtMap`.

One work-around would be to use `DbtVec<Option<u64>>` for the underlying
list of keys and mark deleted keys as `None`. But this would result in
incomplete data deletion when `remove` was used (a `None` element would
remain in underlying `DbtVec`), and it would result in linear instead of
constant running time for the map's `len` and `is_empty`
implementations. Instead of leaving these footguns lying around, I
though it'd be better to restrict functionality. The intended use case
doesn't need `remove` anyway.

Cf. #786.