refactor(wallet): simplify logic for setting initial schema version
If wallet-DB is empty, there is no need to migrate, so we just set
schema version to most recent. Previously this logic was more convoluted
in that it checked that the schema_version was also the singleton's
default value. But setting that default value happened far away from
this check, so it was quite opaque, I think. This makes the check
simple: if sync_label is default digest, wallet database is empty.