refactor(wallet)!: Add cryptographic data to monitored UTXO
Previously this cryptographic data was only available through the
monitored UTXO's membership proofs. With this commit it is moved out
into the structure itself which provides multiple benefits, for example
that the required new mapping `aocl_to_mutxo` mapping can be populated
on the insertion of monitored UTXOs into the database.