refactor!(wallet): Fast index set -> monitored UTXO lookup
Add a new table to the wallet database to allow for fast lookups from
absolute index sets to monitored UTXOs.

This is used to speed up the `WalletState::scan_for_spent_utxos` method
that is called each time the state is updated with a new block.

This method goes from haven an O(N) runtime to having a O(L) runtime,
where N is number of monitored UTXOs, and L is number of inputs in the
applied block.