refactor(sync-loop): Punish unresponsive peers
Whenever the sync loop is active, every 10 seconds that a peer does not respond,
it will be punished.

This mechanic is achieved by tracking two new pieces of information, stored on
the `PeerSyncState`. Both are optional timestamps: `last_response` and
`last_punishment`. The former tracks the last any message from the peer (so that
punishments do not happen before they need to). The second tracks the last time
the peer was punished (so that punishments do not happen too frequently).