feat(sync_loop): Sync from syncing peers
When syncing, the node now asks peers who are also syncing about blocks they can
serve. Upon being asked for a block by height, the client now first checks the
archival state. If the archival state does not have the block, and if the sync
loop is active, then the block request message is passed on to the sync loop.
The sync loop responds with the block if it is present, or with the
synchronization bit mask (AKA coverage) if not.

Caution: this commits adds a new `PeerMessage`: `SyncCoverage`, whose associated
data is the synchronization bit mask indicating which blocks are present and
which are absent.