feat(CLI): Add command `no-resume-sync` to avoid resuming a previous sync
Downloaded but unprocessed blocks are stored in a temporary directory. By
default, if a new sync is spawned it resumes the sync based on the blocks
stored in that directory. The new command allows the user to specify that
no resumption is allowed. If a new sync process is started, it must start
from scratch.