feat: Tolerate reorgs and warn user about undeletable corrupt data
Tolerance for reorgs is already implicitly present: a sync failure will lead to
a new `RequestBlockNotification` which should trigger a whole new sync. In order
for the new sync to avoid using potentially corrupt data, it is important to
avoid resuming the failed sync. By default, the sync loop deletes the temporary
directory holding its state, but this deletion may fail. In this case, inform
the user about deleting it manually.

This commit adds the informational log message (as an error message).

Also: document behavior.