refactor(`SyncProgress`): Change meaning of `SyncProgress`
Previously, `SyncProgress` was a number indicating progress relative to the
interval of missing blocks. However, this turns out to be a very noisy and
unreliable number in the context of sync across forks when LUCA is unknown.
LUCA is generally unknown.

Phrased differently, if the magnitude of the missing interval were known, then
LUCA would be known as well; and finding LUCA is precisely the sync loops job.

So instead, the `SyncProgress` captures the total number of blocks downloaded
relative the total number of blocks on the chain.