refactor(sync_loop): Drop early exit backstop
Previously, when there was no sign of life from peers for too long, the sync
loop would terminate. This escape hatch has been removed.

First, it was designed to recover from an inconsistent state, which should not
be reachable to begin with.

Second, the criteria for early exiting are met in practice falsely when there
is heavy load on the communication channels. (For instance, this can happen
when the main loop is busy processing blocks and each state update takes a
while because there are many UTXOs to keep track of.)