refactor: Move and clarify const for mempool/tx-relay policy
Const `MAX_ANNOUNCEMENT_MESSAGE_SIZE` is does not contribute to block
(in)validity, but is enfored at the peer-to-peer level as a transaction relay
and mempool insertion policy. This commit moves the const definition to the
file where it is used, `peer_loop.rs` and clarifies its purpose in the
docstring.