fix(wallet): Set sync label on wallet-DB recovery if DB is new
If the wallet DB's sync label was never set, that means that the wallet
DB is new. In that case, all possible membership proof recoveries have
been performed when restoring the wallet database from an archival
mutator set. If that's the case, the sync label should be set.

Also: Change logic for when a wallet database is considered new to take
the wallet DB version into account.