feat(archival_state): Add UTXO index
Add a UTXO index for fast wallet syncing over many blocks. This new
index holds information about announcements, and outputs of a block.
Together, this index allows for the efficient syncing of wallets
without having to load all blocks from disk, as reading from the
databases is up to 1000x faster than loading the entire block into
memory and fetching the relevant informatino.

The index does not and should not include addition records in a block,
since all addition records are present as leaves in the archival
mutator set, which is assumed to be part of the state of anyone
maintaining a UTXO index.

The UTXO index maintanence is activated by starting the node with the
`--utxo-index` flag.

This commit adds a `rescan_announced_incoming` method to global state
that allows for the scanning of announced transactions only.