perf: Optionally restrict range of UTXOs restored
Iterate only over new UTXOs, not all UTXOs, in the course of MUTXO MSMP
restoration. Makes syncing faster. But do this only when a flag is set,
so that we can set it while syncing (and default to the full range in all
other cases).

Credit to AllFather team. Docstring modified by Alan.