feat(DbtMap): Add `clear` method
Add `clear` method to `DbtMap` that deletes all entries in the map. Is
added because we want the option of deleting all expected MUTXOs, and we
want a lookup from addition records to expected UTXOs for fast indexing,
and if we clear the list of expected UTXOs, we need to also clear the
lookup map.