feat: Add wallet rescanning endpoints to JSON RPC server
Since the RPC server cannot mutate global state, the implied mutation
must be handled as a message from the RPC server to the main loop.

This function holds a write-lock over global state, so many things will
not work while it is running. Maybe that's OK though, since these are
endpoints that will be called very rarely.