perf: Ignore MSMP recoveries for all but trailing k monitored UTXOs
Modify the optimization of function
`restore_monitored_utxos_from_archival_mutator_set` whereby only the trailing
end of the list of monitored UTXOs is processed if `process_all` is false and
if `new_utxos` contains some non-empty vector.

As a result of @Sword-Smith's new database schema, all necessary data for
fetching a mutator set membership proof for a given UTXO lives on the
`MonitoredUtxo` struct. Consequently, the corresponding data for monitored UTXOs
that do not have a membership proof yet is already available -- and there is no
need to pass this data as an argument. Indeed, the PR that applied that database
schema change already dropped the argument `new_utxos`.

Meanwhile, PR `sync-performance-fixes` used argument `new_utxos` for an
optimization, namely to skip the updates of all monitored UTXOs that already
have some MSMP, even if it is outdated. The relevant information contained by
the argument `new_utxos` is its length (if `Some`), because based on the length
of the tail we know where to set the starting index for the loop.

This PR merges arguments `process_all` and `new_utxos` into one argument,
`ignore_all_but_trailing` of type `Option<usize>`. Set it to `None` for the
default behavior, in which case none of the monitored UTXOs will be ignored.
During syncing, or other cases where we don't care about outdated membership
proofs and only care about populating the `k` trailing new monitored UTXOs with
some membership proofs, set it to `Some(k)`.