refactor: Specify peers as `Multiaddr`s
This commit changes the type of command-line argument `peers` from
`Vec<SocketAddr>` to `Vec<Multiaddr>`, where the second item type is inherited
from libp2p.

https://docs.libp2p.io/concepts/fundamentals/addressing/

Users can specify peers in the multiaddress format or in the legacy format.

The existing peer-to-peer related functions continue to work over the subset of
CLI peers that are convertible to `SocketAddr`s.