refactor: Change type of `instance_id` to `PeerId`
The `instance_id` is a random integer whose main purpose is to determine if you
are connecting to the same peer twice or connecting to yourself. To achieve this
goal it must be set uniformly at random at startup.

The `PeerId` from libp2p is used for the same purpose and additionally links the
peer to a public key. In the words of the documentation,

> As well as serving as a unique identifier for each peer, a Peer ID is a
> verifiable link between a peer and its public cryptographic key.

source: https://docs.libp2p.io/concepts/fundamentals/peers/#peer-id

The trouble is that the legacy parts of the code, notably the `HandshakeData`,
require the `u128` `instance_id`. Therefore, a deterministic map is implemented
and used that sends the `PeerId` to a `u128`.