bench(wallet): Add benchmark of wallet rescanning
Add a benhcmark for the rescanning of incoming UTXOs, both announced and
expected. Very happy with the results here, for the rescanning of 110
blocks, of which 10 contain 10 UTXOs for wallet each. With 200 keys
present in wallet.

This benchmark shows that the rescanning times will likely stay under
one minute for the foreseeable future, for all nodes, including mining
pools and exchanges.

wallet_rescan                            fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ rescan                                              │               │               │               │         │
   ├─ rescan_110_blocks_announced_bench  119.8 ms      │ 168.2 ms      │ 131.6 ms      │ 134.6 ms      │ 10      │ 10
   ├─ rescan_110_blocks_expected_bench   37.08 ms      │ 41.9 ms       │ 37.98 ms      │ 38.37 ms      │ 10      │ 10
   ╰─ rescan_110_blocks_outgoing_bench   843.9 µs      │ 1.552 ms      │ 889.6 µs      │ 1.037 ms      │ 10      │ 10